* 'sat' of https://github.com/amisrs/tutor:
  added moodle pic
  tts
  i forgot what i did
  delete tutorial works
  mark input to number
  mark save is good (doesn't reload fragment anymore)
  marks work
  marks work mostly

# Conflicts:
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/com/amisrs/gavin/tutorhelp/view/NewTutorialDialogFragment.java
#	app/src/main/java/com/amisrs/gavin/tutorhelp/view/StudentDetailsFragment.java
#	app/src/main/java/com/amisrs/gavin/tutorhelp/view/TutorialListActivity.java
#	app/src/main/res/layout/fragment_student_details.xml

+++ app/src/main/java/com/amisrs/gavin/tutorhelp/model/Enrolment.java
@@ -6,9 +6,9 @@ package com.amisrs.gavin.tutorhelp.model;
 public class Enrolment {
     public int studentID;
     public int tutorialID;
-    public int grade;
+    public double grade;
 
-    public Enrolment(int studentID, int tutorialID, int grade) {
+    public Enrolment(int studentID, int tutorialID, double grade) {
         this.studentID = studentID;
         this.tutorialID = tutorialID;
         this.grade = grade;
@@ -30,11 +30,11 @@ public class Enrolment {
         this.tutorialID = tutorialID;
     }
 
-    public int getGrade() {
+    public double getGrade() {
         return grade;
     }
 
-    public void setGrade(int grade) {
+    public void setGrade(double grade) {
         this.grade = grade;
     }
 }
